==> 4-westbrook-analysis.Rmd <==
---
title: "run JAGS"
author: "Ben Letcher and Daniel Hocking"
date: "October 3, 2014"
output: html_document
---


NO, ar(2)? doesn't seem worth it
YES, keep cubic? if not, add q?

DIDN'T WORK. structure on ar(), try RE by dOY

correlation of ar with other variables, airTemplagged1 correlated?

```{r results = 'hide', warning=FALSE, message=FALSE}
library(knitr)
library(rjags)
library(reshape2)
library(ggplot2)
library(ggmcmc) # this uses plyr which conflicts with dplyr and makes group_by not work. make sure to load dplyr after this
library(nlme)
library(devtools)
install_github("Conte-Ecology/conteStreamTemperature")
library(conteStreamTemperature)
#library(beepr)
library(dplyr)
```

```{r load data}
baseDir <- getwd() # works as long as you have the project loaded in RStudio - does not work for kniting

dataInDir <- paste0(baseDir, '/dataIn/')
dataOutDir <- paste0(baseDir, '/dataOut/')
dataLocalDir <- paste0(baseDir, '/localData/')
graphsDir <- paste0(baseDir, '/graphs/')

# load standardized and formatted data created with the 3-statModelPrep.Rmd script
load(paste0(dataLocalDir, 'tempDataSync.RData'))

tempDataSyncS$river <- 'WEST BROOK'
tempDataSyncS$river[ tempDataSyncS$site == 'MAUSGS_WB_MITCHELL' ] <- 'WB MITCHELL' 
tempDataSyncS$river[ tempDataSyncS$site == 'MAUSGS_WB_OBEAR' ] <- 'WB OBEAR' 
tempDataSyncS$river[ tempDataSyncS$site == 'MAUSGS_WB_JIMMY' ] <- 'WB JIMMY' 
tempDataSyncS$riverOrdered <- factor(tempDataSyncS$river,levels=c('WEST BROOK','WB JIMMY','WB MITCHELL','WB OBEAR'), ordered=T)

tempDataSyncS$date <- as.Date(tempDataSyncS$date) # to make indexDeployments() work. the posix dates turn into numeric when they are shifted

# merge in env data to get flow
#load(paste0(dataOutDir, 'etWB.RData'))
#et$dateDate <- as.Date(et$date) # to match tempDataSyncS
#et$riverOrdered <- factor(et$site,levels=c('WEST BROOK','WB JIMMY','WB MITCHELL','WB OBEAR'), ordered=T)

et2 <- read.csv(paste0(dataInDir, 'MAUSGS/west drainage discharge.csv'),header=T)
et2$date <- as.Date(strptime(et2$date2, format="%m/%d/%Y") )
et2$riverOrdered <- factor(et2$river,levels=c('WEST BROOK','WB JIMMY','WB MITCHELL','WB OBEAR'), ordered=T)

tempDataSyncS <- left_join(x=tempDataSyncS, y=et2[,c('riverOrdered','flow','date')], )

# add in correct drainage area  - data from hig res flow lines from kyle 
tempDataSyncS$dA <- 0.55833
tempDataSyncS$dA[tempDataSyncS$riverOrdered == "WB JIMMY"] <- 0.02175
tempDataSyncS$dA[tempDataSyncS$riverOrdered == "WB MITCHELL"] <- 0.02932
tempDataSyncS$dA[tempDataSyncS$riverOrdered == "WB OBEAR"] <- 0.01082

# standardize flow etc by site
tempDataSyncS <- 
  tempDataSyncS %>%
    mutate( flowL = log( flow ) ) %>%
    mutate( sweL = log( swe + 1 ) ) %>%
  group_by(site) %>%
    mutate( flowS = ( flow - mean(flow,na.rm = TRUE) ) / sd(flow,na.rm = TRUE) ) %>%
    mutate( flowLS = ( flowL - mean(flowL,na.rm = TRUE) ) / sd(flowL,na.rm = TRUE) ) %>%
    mutate( sweLS = ( sweL - mean(sweL,na.rm = TRUE) ) / sd(sweL,na.rm = TRUE) ) #%>%
   # mutate( ImpoundmentsAllSqKMS = ( ImpoundmentsAllSqKM - mean(ImpoundmentsAllSqKM,na.rm = TRUE) ) / sd(ImpoundmentsAllSqKM,na.rm = TRUE) ) #%>%
    #mutate( dAS = ( dA - mean(dA,na.rm = TRUE) ) / sd(dA,na.rm = TRUE) )       


tempDataSyncS$swe0 <- ifelse(tempDataSyncS$swe < 0, 0, tempDataSyncS$swe )

# there are 10 flows with NA., set to 0
#tempDataSyncS$flowS[is.na(tempDataSyncS$flowS)] <- 0
tempDataSyncS$flowS[is.na(tempDataSyncS$flowS)] <- 0
tempDataSyncS$flowLS[is.na(tempDataSyncS$flowLS)] <- 0

# doY as an integer  
tempDataSyncS$dOYInt <- as.numeric(strftime(tempDataSyncS$date, format = "%j"))
tempDataSync$dOYS <-  (tempDataSync$dOY-mean(tempDataSync$dOY, na.rm=TRUE)) / sd(tempDataSync$dOY, na.rm=TRUE)

# index for testing trend across years-not used in the end
tempDataSyncS$dOYYear <- ( tempDataSyncS$year - min(tempDataSyncS$year) + 1 ) * 365 +
                           tempDataSyncS$dOYInt - 3000 #3000 to center values

tempDataSyncS <- arrange( tempDataSyncS, riverOrdered,date )

```

```{r dummy site variable}

tempDataSyncS$river0 <- ifelse( tempDataSyncS$riverOrdered=='WEST BROOK',1,0 ) # not really needed
tempDataSyncS$river1 <- ifelse( tempDataSyncS$riverOrdered=='WB JIMMY',1,0 )
tempDataSyncS$river2 <- ifelse( tempDataSyncS$riverOrdered=='WB MITCHELL',1,0 )
tempDataSyncS$river3 <- ifelse( tempDataSyncS$riverOrdered=='WB OBEAR',1,0 )

tempDataSyncS$site0 <- ifelse( tempDataSyncS$site=='MAUSGS_SEC30_DL',1,0 ) # not really needed
tempDataSyncS$site1 <- ifelse( tempDataSyncS$site=='MAUSGS_SEC45_DL',1,0 )
tempDataSyncS$site2 <- ifelse( tempDataSyncS$site=='MAUSGS_SEC6_DL',1,0 )
tempDataSyncS$site3 <- ifelse( tempDataSyncS$site=='MAUSGS_WB_JIMMY',1,0 )
tempDataSyncS$site4 <- ifelse( tempDataSyncS$site=='MAUSGS_WB_MITCHELL',1,0 )
tempDataSyncS$site5 <- ifelse( tempDataSyncS$site=='MAUSGS_WB_OBEAR',1,0 )
tempDataSyncS$site6 <- ifelse( tempDataSyncS$site=='MAUSGS_WEST_BROOK',1,0 )

```

```{r subset data}

tempDataSyncS <- filter(tempDataSyncS
                        , year <= 2013  
                        ,site != 'MAUSGS_SEC_45_DL'
                        ,site != 'MAUSGS_SEC_30_DL'
                        ,riverOrdered != 'WB MITCHELL' | year != 2013 #construction led to funky data
                        #,site == 'MAUSGS_WB_MITCHELL' 
                        )#'MAUSGS_WEST_BROOK')

tempDataSyncS$rowNum <- 1:nrow(tempDataSyncS)

```

```{r raw data graphs}

ggplot(tempDataSyncS,aes(date,temp))+
  geom_point()+
  geom_point(aes(date,airTemp))+
  geom_point(aes(date,flowLS),color='red')+
  facet_wrap(~site)

ggplot(tempDataSyncS,aes(dOY,temp))+
  geom_point(aes(color=factor(year)))+
  geom_line(aes(color=factor(year)))+
  facet_grid(year~site)
  
ggplot(tempDataSyncS,aes(dOY,temp))+
  geom_point(aes(color=factor(year))) +
  facet_grid(year~riverOrdered)

ggplot(et2,aes(date,flow))+
  geom_point()+
  geom_point(aes(date,temp),tempDataSyncS)+
  facet_wrap(~riverOrdered)


```

```{r run JAGS model}

# try moving mean air T?
# trend across years? use derived paramters
# look at correlation btw snow melt last day in spring and spring bp
# get peak day and max value at peak day from RE. any trend?

data.fixed <- data.frame(intercept = 1
                         ,airTemp = tempDataSyncS$airTemp 
                         ,airTempLag1 = tempDataSyncS$airTempLagged1
                         ,airTempLag2 = tempDataSyncS$airTempLagged2
                         
                         ,flow =  tempDataSyncS$flowLS
                         
                         ,airFlow = tempDataSyncS$airTemp * tempDataSyncS$flowLS
#                         ,air1Flow = tempDataSyncS$airTempLagged1 * tempDataSyncS$flowLS
#                         ,air2Flow = tempDataSyncS$airTempLagged2 * tempDataSyncS$flowLS
                         
                         #main river effects
                         ,river1 = tempDataSyncS$river1
                         ,river2 = tempDataSyncS$river2
                         ,river3 = tempDataSyncS$river3
                         
                         #river interaction with air temp
                         ,river1Air = tempDataSyncS$river1 * tempDataSyncS$airTemp
                         ,river2Air = tempDataSyncS$river2 * tempDataSyncS$airTemp
                         ,river3Air = tempDataSyncS$river3 * tempDataSyncS$airTemp
                         
                         # river interaction with flow
        #                 ,river1Flow = tempDataSyncS$river1 * tempDataSyncS$flowLS
        #                 ,river2Flow = tempDataSyncS$river2 * tempDataSyncS$flowLS
        #                 ,river3Flow = tempDataSyncS$river3 * tempDataSyncS$flowLS

                        #  ,imp = tempDataSyncS$ImpoundmentsAllSqKM
                       #   ,dA = tempDataSyncS$dA                                        Very high CI
                      #   ,swe = tempDataSyncS$swe0   
         #                ,sweFlow = tempDataSyncS$swe0 * tempDataSyncS$flowLS

#                          ,dOYYear0 = tempDataSyncS$dOYYear * tempDataSyncS$river0
#                          ,dOYYear1 = tempDataSyncS$dOYYear * tempDataSyncS$river1 
#                          ,dOYYear2 = tempDataSyncS$dOYYear * tempDataSyncS$river2   
#                          ,dOYYear3 = tempDataSyncS$dOYYear * tempDataSyncS$river3
                          ) 

data.random.years <- data.frame(intercept.year = 1, 
                     dOY  = tempDataSyncS$dOY, 
                     dOY2 = tempDataSyncS$dOY^2,
                     dOY3 = tempDataSyncS$dOY^3
                     )

monitor.params <- c("residuals",
            "deviance",
 #           "pD",
            "sigma",
            "B.0",
            "B.year",
            "rho.B.year",
            "mu.year",
            "sigma.b.year",
            "stream.mu",
            "ar1" ,
            "ar1Mean",
            "ar1SD"
            )

coda.tf <- T # currently only works in full for TRUE (using coda.samples)

tempDataSyncS$HUC8 <- NA #Dan changed the function - need this to make existing function work
tempDataSyncS <- indexDeployments( data.frame(tempDataSyncS) )
firstObsRows <- createFirstRows(tempDataSyncS)
evalRows <- createEvalRows(tempDataSyncS)

system.time(M.wb <- modelRegionalTempWB(tempDataSyncS, firstObsRows, evalRows, data.fixed=data.fixed, data.random.years=data.random.years, n.burn = 1000, n.it = 5000, n.thin = 5, nc = 3, coda = coda.tf, param.list = monitor.params))

save(M.wb, file = paste0(dataLocalDir, "westbrook-mcmc.RData"))

rm(ggs.wb)
ggs.wb <- ggs(M.wb)
gc() # get rid of ggs mess - may not be necessary on a pc


  # add mean residuals and predicted values from the jags run
  # this replaces Dan's loop. It's a lot more code but a lot faster.
  tempDataSyncS <- addStreamMuResid(M.wb,tempDataSyncS)

huc.mat <- as.matrix(M.wb)

  rmseBy <- 
    tempDataSyncS %>%
    group_by( year,riverOrdered ) %>%
    summarise( rmse = rmse( resid.wb ) ) %>%
    dcast(year~riverOrdered)

means <- ggs.wb %>%
  group_by(Parameter)%>%
  summarise(mean=mean(value),
            sd=sd(value),
            qLo=quantile(value,probs=c(0.025)),
            qHi=quantile(value,probs=c(0.975)),
            qLo25=quantile(value,probs=c(0.25)),
            qHi75=quantile(value,probs=c(0.75))
            )

save(rmseBy,means,file=paste0(dataLocalDir,'means.RData'))
save( tempDataSyncS,  file = paste0(dataLocalDir,'tempDataSyncSUsed.RData') )
save( ggs.wb,  file = paste0(dataLocalDir,'ggs_wb.RData') )

```

## Evaluate MCMC Iterations
```{r check model convergence and mixing}

# this saves a pdf file
ggmcmc(ggs.wb, file = "ggmcmc-wb.pdf", family = "B.0", plot = c("ggs_density",  "ggs_traceplot", "ggs_running", "ggs_compare_partial", "ggs_autocorrelation", "ggs_crosscorrelation", "ggs_Rhat", "ggs_geweke", "ggs_caterpillar"))
dev.off()

ggmcmc(ggs.wb, file = "ggmcmc-wb.pdf", family = "B.year", plot = c("ggs_density",  "ggs_traceplot", "ggs_running", "ggs_compare_partial", "ggs_autocorrelation", "ggs_crosscorrelation", "ggs_Rhat", "ggs_geweke", "ggs_caterpillar"))
dev.off()

ggs_traceplot(ggs.wb[ggs.wb$Parameter %in% c('B.0[1]','B.0[2]','B.0[3]','B.0[4]','B.0[5]','B.0[6]'),], family = "B.0")
ggs_traceplot(ggs.wb[ggs.wb$Parameter %in% c('B.0[14]','B.0[15]','B.0[16]','B.0[17]','B.0[18]','B.0[19]'),], family = "B.0")

ggs_traceplot(ggs.wb[ggs.wb$Parameter %in% c('ar1[1]','ar1[2]','ar1[3]','ar1[4]','ar1Mean','ar1SD'),])

# these graphs are the evidence for rHat < 1.01
ggs_Rhat(ggs.wb, family = "B.0")
ggs_Rhat(ggs.wb, family = "B.year")
ggs_Rhat(ggs.wb, family = "ar1",scaling=1.1)


ggs_density(ggs.wb, family = "B.0")
ggs_crosscorrelation(ggs.wb, family = "B.0")

ggs_density(ggs.wb, family="B.year")
ggs_traceplot(ggs.wb, family = "B.year")

ggs_traceplot(ggs.wb, family = "ar1")
ggs_density(ggs.wb, family="ar1")

mean(ggs.wb[ggs.wb$Parameter == 'ar1[2]','value'])
mean(ggs.wb[ggs.wb$Parameter == 'ar1Mean','value'])



```

## Check model fit
```{r check model fit}
#------ Check residual patterns------
# Observed vs Residuals
g <- ggplot(tempDataSyncS, aes(temp, resid.wb)) 
g + geom_point(aes(color=factor(year)),alpha = 0.3) + geom_smooth() + theme_bw()


#observed vs. predicted
  ggplot(tempDataSyncS
         #[tempDataSyncS$year>2006,]
         , aes(temp, pred.wb)) +
    geom_point(alpha = 0.3) + 
    geom_abline(intercept=0,slope=1) + 
    theme_bw() +
    facet_grid(year~riverOrdered)

yearPlot <- 2007
ggplot(tempDataSyncS[
                     tempDataSyncS$year %in% yearPlot
              #       &  tempDataSyncS$site == 'MAUSGS_WB_MITCHELL'
                     &  tempDataSyncS$river == 'WEST BROOK'
                     ,], aes(dOY, pred.wb)) + 
  geom_point(aes(dOY,temp),color='blue')+
  geom_line(aes(dOY,temp),color='blue')+
  geom_line(aes(dOYS ,airTemp),color='black', data=tempDataSync[tempDataSync$year %in% yearPlot,])+
  geom_point(aes(dOYS ,airTemp),color='black', data=tempDataSync[tempDataSync$year %in% yearPlot,])+
 # geom_line(aes(dOY,pred),data=predCubic[predCubic$year==yearPlot-2005,])+
 #  geom_line(aes(dOY,pred,color=factor(year)),data=predCubic)+
  geom_point(color='red') +
  facet_wrap(~riverOrdered) +
xlim(-1.2,-.6)  



library(hexbin)
g + stat_binhex(bins = 100) + geom_smooth() + theme_bw()

g + geom_point() + geom_density2d()

# Residuals by random year
b <- ggplot(tempDataSyncS, aes(x = factor(year), y = resid.wb))
b + geom_boxplot() + coord_flip()

# Residuals by covariates
ggplot(tempDataSyncS, aes(airTemp, resid.wb)) + geom_point(alpha = 0.5) + geom_smooth() + theme_bw() + facet_wrap(~site)

ggplot(tempDataSyncS, aes(swe0*flowS, resid.wb)) + geom_point(alpha = 0.5) + geom_smooth() + theme_bw() + facet_wrap(~site)

ggplot(tempDataSyncS[tempDataSyncS$year==2010,], aes(dOY, resid.wb)) + geom_point(alpha = 0.5) + geom_smooth() + theme_bw()+ facet_grid(site~.)

ggplot(tempDataSyncS, aes(dOY^2, resid.wb)) + geom_point(alpha = 0.5) + geom_smooth() + theme_bw()

# correlation in MCMC for residuals
acf(huc.mat[ , "residuals[1]"], lag.max = 40, xlab = "Lag", ylab = "Correlation", main = "")
acf(huc.mat[ , "residuals[2]"], lag.max = 40, xlab = "Lag", ylab = "Correlation", main = "")

# correlation in residuals over time 
acf(tempDataSyncS$resid.wb, lag.max = 100, xlab = "Lag", ylab = "Correlation", main = "") # 

```

```{r cross-validation - many scenarios}
#assumes code has been run up to the model call in 'run Jags model' chunk

n.burn <- 100
n.it <- 50
n.thin <- 2

years <- data.frame(year=unique(tempDataSyncS$year)) %>%
           arrange(.,year)

#years=data.frame(year=1997:2000)

ggObsPred <- ggCubic <- list()
overlap <- rep(NA,5)
rm(rmseByOut,rmseByOutAll)
rmseByOut <- rmseByOutAll <- predictSummary <- NA

# scenarios to test. set up  different sets of days to keep
nDays <- 30 # num of days to keep

loDOY  <- round(quantile(tempDataSyncS$dOYInt,0.15))
mDOY  <- round(quantile(tempDataSyncS$dOYInt,0.5))
hiDOY  <- round(quantile(tempDataSyncS$dOYInt,0.85))

rm(dOYKeep1); dOYKeep1 <- matrix(NA,28,nDays*10)
# add 1 row at end to hold as all NAs. this is the case off no obs for a given year
iD <- 0
for( iDay in seq(70,310,20) ){
  iD <- iD+1
  dOYKeep1[iD,1:length(iDay:(iDay+nDays-1) )] <- iDay:(iDay+nDays-1) 
  #dOYKeep1[2,1:length(270:(270+nDays-1))] <- 270:(270+nDays-1)
}

#spread from the middle in both directions
dOYKeep1[14,1:length((mDOY-nDays/2):(mDOY+nDays/2-1))] <- (mDOY-nDays/2):(mDOY+nDays/2-1)
dOYKeep1[15,1:length((mDOY-nDays*1):(mDOY+nDays*1-1))] <- (mDOY-nDays*1):(mDOY+nDays*1-1)
dOYKeep1[16,1:length((mDOY-nDays*1):(mDOY+nDays*1.5-1))] <- (mDOY-nDays*1):(mDOY+nDays*1.5-1)
dOYKeep1[17,1:length((mDOY-nDays*2):(mDOY+nDays*2-1))] <- (mDOY-nDays*2):(mDOY+nDays*2-1)
dOYKeep1[18,1:length((mDOY-nDays*2.5):(mDOY+nDays*2.5-1))] <- (mDOY-nDays*2.5):(mDOY+nDays*2.5-1)
dOYKeep1[19,1:length((mDOY-nDays*3):(mDOY+nDays*3-1))] <- (mDOY-nDays*3):(mDOY+nDays*3-1)
dOYKeep1[20,1:length((mDOY-nDays*3.5):(mDOY+nDays*3.5-1))] <- (mDOY-nDays*3.5):(mDOY+nDays*3.5-1)
dOYKeep1[21,1:length((mDOY-nDays*4):(mDOY+nDays*4-1))] <- (mDOY-nDays*4):(mDOY+nDays*4-1)
dOYKeep1[22,1:length((mDOY-nDays*4.5):(mDOY+nDays*4.5-1))] <- (mDOY-nDays*4.5):(mDOY+nDays*4.5-1)

# random picks
dOYKeep1[23,1:length(sample(loDOY:hiDOY,nDays))] <- sample(loDOY:hiDOY,nDays)
dOYKeep1[24,1:length(sample(loDOY:hiDOY,nDays))] <- sample(loDOY:hiDOY,nDays)
dOYKeep1[25,1:length(sample(loDOY:hiDOY,nDays))] <- sample(loDOY:hiDOY,nDays)
dOYKeep1[26,1:length(sample(loDOY:hiDOY,nDays))] <- sample(loDOY:hiDOY,nDays)
dOYKeep1[27,1:length(sample(loDOY:hiDOY,nDays))] <- sample(loDOY:hiDOY,nDays)

dOYKeep1[28,] <- NA # just as a reminder. No data for the year case


rivs <- list()
rivs[[1]] <- "WEST BROOK"
rivs[[2]] <- unique(tempDataSyncS$riverOrdered)

inc <- 0

# loop through either excluing just the west brook or all rivers in a given year/scenario
for ( rivsIndex in 1:2 ){

# loop through the scenarios
for ( k in 1:nrow( dOYKeep1 ) ){
#for ( k in c(12,18,26,28) ){

dOYKeep <- dOYKeep1[ k, ] 
dOYKeep <- dOYKeep[ !is.na( dOYKeep ) ]
minDOY <- min(dOYKeep)
maxDOY <- max(dOYKeep)

# loop over years
for ( i in 1:( nrow(years) + 0 ) ){
# for ( i in 12 ){
 
  inc <- inc + 1
  y <- years$year[i]
  #print(c(inc,k,i,y))

  dat <- tempDataSyncS
  dat$temp[ dat$year %in% y 
            & dat$dOYInt %!in% dOYKeep
            & tempDataSyncS$riverOrdered %in% rivs[[rivsIndex]]
          ] <- NA
    
  dat <- indexDeployments( data.frame(dat) )
  firstObsRows <- createFirstRows(dat)
  evalRows <- createEvalRows(dat)
  
  dat$eval <- dat$rowNum %in% evalRows
  dat$first <- dat$rowNum %in% firstObsRows
  
  overlap1 <-  
    dat %>%
      group_by( riverOrdered ) %>%
      filter( year == y ) %>%
      mutate( #numD = length( intersect( dOYInt , dOYKeep ) ),
              numD = sum( !is.na(temp) ),
              propD = numD / length( dOYInt ),
              rivsIndex=rivsIndex,
              scenario = k,
              year = y) %>%
      select( numD,propD,scenario,year,rivsIndex ) %>%
      distinct(riverOrdered,numD)
  
  overlap <- rbind(overlap,overlap1)

  print(c(inc,rivsIndex,k,i,y,overlap1$propD))
  
  #Don't do any estimation if there are no firstObsRows for this year
  if( length(intersect( dat[ dat$year == y,'rowNum' ],firstObsRows ) ) > 0 ){
    
    print( system.time(M.wb <- modelRegionalTempWB(dat, firstObsRows, evalRows,
                                          data.fixed=data.fixed, 
                                          data.random.years=data.random.years, 
                                          n.burn = n.burn, n.it = n.it, n.thin = n.thin, nc = 3,
                                          coda = coda.tf, param.list = monitor.params)) )
  
    save(M.wb, file = paste0(dataLocalDir, "westbrook-mcmcValidate_",rivsIndex,"_",k,"_",y,".RData"))
  
  # add mean residuals and predicted values from the jags run
  # this replaces Dan's loop. It's a lot more code but a lot faster.
  tempDataSyncS <- addStreamMuResid(M.wb,tempDataSyncS)

  # residuals based on predicted temps and acutal observed temps ( including the temps left out )
  tempDataSyncS$resid <- tempDataSyncS$temp - tempDataSyncS$pred.wb

  #rm(rmse)
  print(c( rmse(tempDataSyncS$resid),rmse(tempDataSyncS$resid.wb)) )
  
  rmseBy <- 
    tempDataSyncS %>%
    group_by( year,riverOrdered ) %>%
    summarise( rmse = rmse( resid ) ) %>%
    mutate( rivsIndex=rivsIndex,
            scenario = k,
            yearIn = years$year[i],
            minDOY = minDOY,
            maxDOY = maxDOY ) %>%
    left_join( overlap1 )

  print( rmseBy %>% dcast( year~riverOrdered, value.var="rmse" ) )  

  rmseByOutAll <- rbind(rmseByOutAll,rmseBy)

  predictSummary <- rbind( predictSummary, data.frame( r=tempDataSyncS$river, year=tempDataSyncS$year, dOYInt = tempDataSyncS$dOYInt, rivsIndex=rivsIndex, scenario = k, yearOut = y, minDOY=minDOY, maxDOY=maxDOY, temp = tempDataSyncS$temp, pred = tempDataSyncS$pred.wb, resid = tempDataSyncS$resid ) )

#   means <- ggs(M.wb) %>%
#     group_by(Parameter)%>%
#     summarise(mean=mean(value),
#               sd=sd(value),
#               qLo=quantile(value,probs=c(0.025)),
#               qHi=quantile(value,probs=c(0.975))
#               )
# 
#   pC <- predCubic(means)
#   ggCubic[[inc]] <- ggplot(pC$predCubic, aes(dOY,pred)) + geom_line(aes(color=factor(year)),size=1.25)
#    
#   print( ggCubic[[inc]] )
#   
                           
  ggObsPred[[inc]] <- 
  ggplot(tempDataSyncS
        [tempDataSyncS$year %in% y,]
         , aes(dOY, pred.wb)) + 
    geom_point(aes(dOY,temp),color='blue')+
    geom_line(aes(dOY,temp),color='blue')+
    geom_line(aes(dOYS ,airTemp),color='black', data=tempDataSync[tempDataSync$year %in% y,])+
    geom_point(aes(dOYS ,airTemp),color='black', data=tempDataSync[tempDataSync$year %in% y,])+
    geom_point(color='red') +
    #geom_point( aes(dOY,1), data=data.frame(dOY=dOYKeep))+
    ggtitle(paste(inc,'rivsIndex = ',rivsIndex,' Scenario = ',k,' Year =',y)) +
    facet_wrap(~riverOrdered)


    if( nrow(dat[dat$year %in% y & dat$rowNum %in% evalRows,] ) > 0 ){
        ggObsPred[[inc]] <-
          ggObsPred[[inc]] +
          geom_point( aes(dOY,0), color='darkgreen',data=dat[dat$year %in% y & dat$rowNum %in% evalRows,]  )
    }    
    if( nrow(dat[dat$year %in% y & dat$rowNum %in% firstObsRows,] ) > 0 ){
        ggObsPred[[inc]] <-
          ggObsPred[[inc]] +
          geom_point( aes(dOY,0.5), color='darkred',data=dat[dat$year %in% y & dat$rowNum %in% firstObsRows,]  )
    }     

  print( ggObsPred[[inc]] )

  rmseByOut <- rbind( rmseByOut, rmseBy[ rmseBy$year == y, ] )

  #beep(10) this messes up the looping
  } # bracket for if statement checking intersection of evalRows and rowNum by year
 } # year bracket 
} #k bracket
} #river bracket
predictSummary <- na.omit(predictSummary)

save( dOYKeep1,rmseByOut,rmseByOutAll,predictSummary,overlap,ggObsPred,ggCubic, file = paste0(dataLocalDir,'validation.RData'))

#beep()


#########################################################################################
#Base case using all data for comparison
# this gets run separately from the looping above

  tempDataSyncS <- indexDeployments( data.frame(tempDataSyncS) )
  firstObsRows <- createFirstRows(tempDataSyncS)
  evalRows <- createEvalRows(tempDataSyncS)

  print( system.time(M.wb <- modelRegionalTempWB(tempDataSyncS, firstObsRows, evalRows,
                                          data.fixed=data.fixed, 
                                          data.random.years=data.random.years, 
                                          n.burn = n.burn, n.it = n.it, n.thin = n.thin, nc = 3,
                                          coda = coda.tf, param.list = monitor.params)) )
  
  save(M.wb, file = paste0(dataLocalDir, "westbrook-mcmcValidateAllData.RData"))
  
  # add mean residuals and predicted values from the jags run
  # this replaces Dan's loop. It's a lot more code but a lot faster.
  tempDataSyncS <- addStreamMuResid(M.wb,tempDataSyncS)

  # residuals based on predicted temps and acutal observed temps ( including the temps left out )
  tempDataSyncS$resid <- tempDataSyncS$temp - tempDataSyncS$pred.wb

  #rm(rmse)
  print(c( rmse(tempDataSyncS$resid),rmse(tempDataSyncS$resid.wb)) )
  
  rmseByComplete <- 
    tempDataSyncS %>%
    group_by( year,riverOrdered ) %>%
    summarise( rmseComplete = rmse( resid ) )# %>%
  #  mutate( minDOY = minDOY,
  #          maxDOY = maxDOY)

  print( rmseByComplete %>% dcast( year~riverOrdered, value.var="rmseComplete" ) ) 

#beep(10)
#########################################################################################

```

```{r striaght up cross-validation - leave out p randomly across years}

# ugly repeat of code, but mostly repeating above with a simpler loop

n.burn <- 100
n.it <- 50
n.thin <- 2

rm(rmseByOutAllCV)
 rmseByOutAllCV <-  NA

# loop over propOut
propOut <- c(0,0.05,seq(0.1,0.9,0.1),0.95,0.99)
for (p in propOut){
 # loop over replicates
 for ( i in 1:10 ){

  dat <- tempDataSyncS
  
  dat$tempOrig <- dat$temp
  dat$keep <- runif(nrow(dat)) > p #0.3  
  if ( p>0 ) dat$temp <- ifelse( dat$keep, dat$temp, NA )  #set p of temp values to NA for non-training runs
   
  dat <- indexDeployments( data.frame(dat) )
  firstObsRows <- createFirstRows(dat)
  evalRows <- createEvalRows(dat)
  
  dat$eval <- dat$rowNum %in% evalRows
  dat$first <- dat$rowNum %in% firstObsRows

  print(c(p,i,table(is.na(dat$temp))))
  
    print( system.time(M.wb <- modelRegionalTempWB(dat, firstObsRows, evalRows,
                                          data.fixed=data.fixed, 
                                          data.random.years=data.random.years, 
                                          n.burn = n.burn, n.it = n.it, n.thin = n.thin, nc = 3,
                                          coda = coda.tf, param.list = monitor.params)) )
  
    save(M.wb, file = paste0(dataLocalDir, "westbrook-mcmcValidateCV_",p,i,".RData"))
  
  # add mean residuals and predicted values from the jags run
  # this replaces Dan's loop. It's a lot more code but a lot faster.
  tempDataSyncS <- addStreamMuResid(M.wb,tempDataSyncS)

  # residuals based on predicted temps and acutal observed temps ( including the temps left out )
  tempDataSyncS$resid <- tempDataSyncS$temp - tempDataSyncS$pred.wb

  tempDataSyncS$keep <- dat$keep
  
  rmseByKeep <- tempDataSyncS %>% 
                  group_by( keep ) %>% 
                  summarise( rmse = rmse( resid ), n=n() )
  
  #rm(rmse)
  print(c( rmse(tempDataSyncS$resid),rmse(tempDataSyncS$resid.wb),rmseByKeep) )

  rmseByOutAllCV <- rbind(rmseByOutAllCV,cbind(propOut=p,rmseAll=rmse(tempDataSyncS$resid),rmseByKeep))

  } # rep bracket 

summ <- rmseByOutAllCV %>% 
          group_by( keep,propOut ) %>% 
            summarise( mean=mean( rmse ), sd=sd( rmse ),
                       loCI= quantile(rmse,probs=0.025,na.rm=T),
                       hiCI= quantile(rmse,probs=0.0975,na.rm=T),
                       meanN=mean(n) )


#predictSummary <- na.omit(predictSummary)

save( rmseByOutAllCV,summ, file = paste0(dataLocalDir,'crossValidationPropOut.RData'))

#print(ggplot(summ,aes(propOut,mean,color=keep)) + geom_point())
#print(ggplot(rmseByOutAllCV,aes(propOut,rmse,color=keep)) + geom_point())

} #proportion of data bracket






tempDataSyncS %>% select(keep,temp,resid,pred.wb) %>% head(20)
dat %>% select(keep,temp,resid,pred.wb) %>% head(20)


y=2011
  ggplot(tempDataSyncS
        [tempDataSyncS$year %in% y,]
         , aes(dOY, pred.wb)) + 
    geom_point(aes(dOY,temp),color='blue')+
    geom_line(aes(dOY,temp),color='blue')+
    geom_line(aes(dOYS ,airTemp),color='black', data=tempDataSync[tempDataSync$year %in% y,])+
    geom_point(aes(dOYS ,airTemp),color='black', data=tempDataSync[tempDataSync$year %in% y,])+
    geom_point(aes(color=keep))+#'red') +
    #geom_point( aes(dOY,1), data=data.frame(dOY=dOYKeep))+
#    ggtitle(paste(inc,'rivsIndex = ',rivsIndex,' Scenario = ',k,' Year =',y)) +
    facet_wrap(~riverOrdered)
```



```{r deviance from base for rmse - old}
# Look at deviance in rmse's from base case

r <- left_join(rmseByOutAll,rmseByComplete)
r$diff <- r$rmse - r$rmseComplete
r$incompleteYear <- r$year == r$yearIn
r$incompleteIndex <- paste(r$minDOY,r$maxDOY,sep='_')
r$dOYDiff <- r$maxDOY - r$minDOY

r$s <- ifelse( r$scenario %in% 1:13, '30days',
       ifelse( r$scenario %in% 14:22, 'middle',
       ifelse( r$scenario %in% 23:27, '30random',
       ifelse( r$scenario %in% 28, '0days', NA)
       )))


save(r, file = paste0(dataLocalDir, "r.RData"))

ggplot(na.omit(r[r$incompleteYear & !is.na(r$riverOrdered),]), aes(year,rmse, color=factor(scenario))) +
  geom_point() + 
  geom_line() +
  geom_point(aes(year,rmseComplete),color='black') +
  facet_grid(rivsIndex~riverOrdered)

ggplot(na.omit(r[r$incompleteYear,]), aes(year,diff, color=factor(scenario))) +
  geom_point() + 
  geom_line() +
  facet_grid(rivsIndex~riverOrdered)

r[r$incompleteYear,] %>%
  group_by( scenario, rivsIndex ) %>%
  summarise( mean = mean( diff, na.rm=T ),
             sd   =   sd( diff, na.rm=T ),
             n    =    n(  ) )  %>%
  ggplot(.,aes(scenario,mean,color=factor(rivsIndex))) +
    geom_point() + geom_line()

r[r$incompleteYear,] %>%
  group_by( riverOrdered,scenario,rivsIndex ) %>%
  summarise( mean = mean( diff, na.rm=T ),
             sd   =   sd( diff, na.rm=T ),
             n    =    n(  ) ) %>%
  ggplot(.,aes(scenario,mean,color=factor(riverOrdered))) +
    geom_point() + geom_line() +
    facet_wrap(~rivsIndex)

r[r$incompleteYear,] %>%
  group_by( year,scenario ) %>%
  summarise( mean = mean( diff, na.rm=T ),
             sd   =   sd( diff, na.rm=T ),
             n    =    n(  ) ) %>%
  ggplot(.,aes(scenario,mean,color=factor(year))) +
    geom_point() + geom_line()

```

```{r deviance from base for pred}
# Look at deviance in pred's from base case
# not updated to include rivsIndex

nRuns <- ( nrow(predictSummary)/nrow(tempDataSyncS) )
predictSummary$rowNum <- rep(1:7113, nRuns)

predictSummary <- left_join(predictSummary, tempDataSyncS[,c('rowNum','year','riverOrdered','dOYInt','pred.wb')], by='rowNum')
predictSummary$diff <- predictSummary$pred - predictSummary$pred.wb

predictSummary$s <- ifelse( predictSummary$scenario %in% 14:22, 'middle',NA) 

ggS <- ggSByYear <- list()
ii=0; iii <- 0
for(rr in c('WEST BROOK', 'WB JIMMY')){
  for(yy in 1997:2012){
    for( ss in c('middle')){
      dat <- predictSummary[predictSummary$year==predictSummary$yearOut &
                            predictSummary$riverOrdered == rr &
                            predictSummary$year == yy &
                            predictSummary$s == ss,]
      dat <- na.omit(dat)
      
      if( nrow( dat ) > 0 ) {
      
      ii=ii+1
      print(c(ii,rr,yy,ss))
ggS[[ii]] <- 
ggplot(dat, aes(dOYInt,diff))+
  geom_point(aes(color=factor(scenario))) +
  geom_line(aes(color=factor(scenario))) +
  scale_colour_hue('clarity') + #h=c(0, 360)) +
  ggtitle(paste('River = ',rr,'  Scenario = ',ss,'  Year =',yy)) 
   }
  }

iii=iii+1
ggSByYear[[iii]] <- 
  ggplot(dat, aes(dOYInt, pred.wb)) + 
    geom_point(aes(dOYInt,pred),color='blue')+
    geom_line(aes(dOYInt,pred),color='blue')+
    geom_point(color='red') +
    ggtitle(paste('River = ',rr,'  Year =',yy)) +
    facet_wrap(~scenario)
 }
}  
  
rmseByOut[rmseByOut$yearIn==2001,]



```

```{r ending info}

print(sessionInfo())

```
