==> 5-westbrook graphs and tables.Rmd <==
cross-validation
rmse by river,year
```{r libraries}
library(ggplot2)
library(ggmcmc) # this uses plyr which conflicts with dplyr and makes group_by not work. make sure to load dplyr after this
library(dplyr)
library(tidyr)
library(gridExtra)
library(gam)
library(gridExtra)
library(conteStreamTemperature)
```



```{r load data}
baseDir <- getwd() # works as long as you have the project loaded in RStudio - does not work for kniting

dataInDir <- paste0(baseDir, '/dataIn/')
dataOutDir <- paste0(baseDir, '/dataOut/')
dataLocalDir <- paste0(baseDir, '/localData/')
graphsDir <- paste0(baseDir, '/graphs/')
codeDir <- paste0(baseDir, '/code/')

load(paste0(dataLocalDir,'means.RData'))
load(paste0(dataLocalDir,'tempDataSyncSUsed.RData'))
#load(paste0(dataLocalDir,'tempDataSync.RData'))

```

```{r table of obs}
ftable(tempDataSync$year,tempDataSync$site)
ftable(tempDataSyncS$year,tempDataSyncS$riverOrdered)

#paste into excel and manipulate table
```


```{r daymet airTemp vs WB airTemp}

load(paste0(dataInDir,'MAUSGS/westbrookAirTempComparison_Paper.RData'))
load(paste0(dataOutDir, 'springFallBreakpoints.RData'))

airTempComparison <- left_join( airTempComparison,springFallBPs[springFallBPs$site=='MAUSGS_WEST_BROOK',],by='year')

# Clip to syncronized season
airTempComparison <- filter(airTempComparison, dOY >= finalSpringBP & dOY <= finalFallBP)
airTempComparison$yearF <- factor(airTempComparison$year)

summary(lm(meanAirTemp_WB~meanAirTemp_Day, data=airTempComparison))

ggAirTempComparison <- 
ggplot(airTempComparison,aes(meanAirTemp_WB,meanAirTemp_Day))+
  geom_point()+
  geom_abline(intercept=0,slope=1) +
  geom_smooth(method='lm') +
  scale_x_continuous(expression(paste("Air temperature from WB ( ",degree, "C)", sep = ""))) +
  scale_y_continuous(expression(paste("Air temperature from dayMet ( ",degree, "C)", sep = ""))) +
  theme_bw(base_size=25)+
  facet_wrap(~year)

ggsave( file=paste0(graphsDir,'ggAirTempComparison.png'), plot=ggAirTempComparison, dpi=600 , width=6,height=5, units='in', scale=2 )

summary( lm( meanAirTemp_Day ~ meanAirTemp_WB*factor(year), data = airTempComparison ) )

ggplot(airTempComparison,aes(dOY,meanAirTemp_Day))+
  geom_point()+
  geom_line()+
  geom_point(aes(dOY,meanAirTemp_WB),color='red') +
    geom_line(aes(dOY,meanAirTemp_WB),color='red') +
  scale_x_continuous(expression(paste("Air temperature from WB ( ",degree, "C)", sep = ""))) +
  scale_y_continuous(expression(paste("Air temperature from dayMet ( ",degree, "C)", sep = ""))) +
  theme_bw(base_size=25)+
  facet_wrap(~year)

#max air temp
ggplot(airTempComparison,aes(dOY,maxAirTemp_Day))+
  geom_point()+
  geom_line()+
  geom_point(aes(dOY,maxAirTemp_WB),color='red') +
    geom_line(aes(dOY,maxAirTemp_WB),color='red') +
  scale_x_continuous(expression(paste("Air temperature from WB ( ",degree, "C)", sep = ""))) +
  scale_y_continuous(expression(paste("Air temperature from dayMet ( ",degree, "C)", sep = ""))) +
  theme_bw(base_size=25)+
  facet_wrap(~year)


```


```{r raw data graphs}
# raw unstandardized are in tempDataSync


###################################################
# graph with tempIndex and breaks for illustration
# tempIndex is in e, not in tempDataSyncS

# to get quantiles
load(paste0(dataOutDir, 'springFallBreakpoints.RData'))
#has temp, airTemp, and tempIndex
load(paste0(dataLocalDir,'e.RData'))

e$river <- 'WEST BROOK'
e$river[ e$site == 'MAUSGS_WB_MITCHELL' ] <- 'WB MITCHELL' 
e$river[ e$site == 'MAUSGS_WB_OBEAR' ] <- 'WB OBEAR' 
e$river[ e$site == 'MAUSGS_WB_JIMMY' ] <- 'WB JIMMY' 
e$riverOrdered <- factor(e$river,levels=c('WEST BROOK','WB JIMMY','WB MITCHELL','WB OBEAR'), ordered=T)
e$month <- as.numeric(strftime(e$date, "%m"))

ggplot(e[e$river=='WEST BROOK',], aes(dOY,temp)) + geom_line() + facet_wrap(~year)
ggplot(e, aes(dOY,temp)) + geom_line() + geom_line(aes(dOY,airTemp),color='red') +
  geom_line(aes(dOY,prcp),color='darkgreen') + facet_grid(river~year)

#ggplot(e, aes(factor(month),temp)) + geom_violin() + facet_grid(year~river)

tempDataSyncS$month <- as.numeric(strftime(tempDataSyncS$date, "%m"))
tempDataSyncS$month <- factor(tempDataSyncS$month,levels=1:12,ordered=T)
meanByMonthRiverYear <- tempDataSyncS %>% group_by(month,riverOrdered,year) %>% 
  summarize(meanByMonthRiverYear=mean(temp),nMonthRiverYear=n())
meanByMonthRiver <- tempDataSyncS %>% group_by(month,riverOrdered) %>% 
  summarize(meanByMonthRiver=mean(temp),nMonthRiver=n())
meanByMonth <- tempDataSyncS %>% group_by(month) %>% 
  summarize(meanByMonth=mean(temp),nMonth=n())

tempDataSyncS <- left_join(tempDataSyncS,meanByMonthRiverYear,by=c("riverOrdered", "year", "month"))
tempDataSyncS <- left_join(tempDataSyncS,meanByMonthRiver,by=c("riverOrdered", "month"))
tempDataSyncS <- left_join(tempDataSyncS,meanByMonth,by=c("month"))

tempDataSyncS$month <- factor(tempDataSyncS$month,levels=1:12,ordered=T)

riverLabellerWB <- function(value){
  value <- as.character(value)
  #if (var=="river") { 
    value[value=="WEST BROOK"] <- "WB"
    value[value=="WB JIMMY"] <- "OL"
    value[value=="WB MITCHELL"] <- "OS"
    value[value=="WB OBEAR"] <- "IL"
  #}
  return(value)
}

tempDataSyncS$riverMS <- riverLabellerWB(tempDataSyncS$riverOrdered)
tempDataSyncS$riverMS <- factor(tempDataSyncS$riverMS,levels=c("WB","OL","OS","IL"), ordered=T)



#tried using meanByMonthRiver directly in a separate geom_point(aes(,,meanByMonthRiver)), but ggplot reorder the factors when months in meanByMonthRiver were present in meanByMonthRiver but not in tempDatSyncS. Ended up just merging meanByMonthRiver into tempDataSycnS
ggV <- list()
for(r in unique(tempDataSyncS$riverMS)){
  ggV[[r]] <- 

# adding in '.x' to the variables just to get the figure done for resubmission    
   ggplot(tempDataSyncS[tempDataSyncS$riverMS==r & tempDataSyncS$nMonthRiverYear.x >= 20 & tempDataSyncS$year >=2003,], aes((month),temp)) + 
    geom_violin() + 
    geom_point(aes((month),meanByMonth.x),shape=16,size=2) +
    geom_point(aes((month),meanByMonthRiver.x),shape=1,size=3) +
    geom_point(aes((month),meanByMonthRiverYear.x),shape=95,size=5) +
    scale_x_discrete('Month') +
    scale_y_continuous("Temperature (C)") +
    theme_bw(base_size=10)+
    ggtitle(r) +
    theme(plot.title = element_text(size=10),
          strip.text = element_text(size=8)) +
    facet_wrap(~year)

  print(ggV[[r]])
  #  ggsave( file=paste0(graphsDir,'ggViolin_',r,'.png'), plot=ggV[[r]], dpi=600 , width=6,height=5, units='in', scale=2 )

}
ggViolin <- arrangeGrob(ggV[[4]],ggV[[1]],ggV[[2]],ggV[[3]], ncol=2)#, widths=c(5,1.5))  

  ggsave( file=paste0(graphsDir,'ggViolin.png'), plot=ggViolin, dpi=600 , width=9,height=5, units='in', scale=2 )



ggplot(e, aes(dOY,temp)) + geom_line(aes(color=factor(year))) + coord_polar() + facet_wrap(~river)

###

labels1 <- data.frame(
  dOY = c(345,345),
  temp = c(24, 24),
  text = c("a", "b"),
  year = c(2003,2004))

ggRawWB <- 
ggplot(e[e$riverOrdered == 'WEST BROOK' & e$year %in% 2003:2004,],aes(dOY,temp))+
  geom_line(size=0.5)+
  geom_line(aes(dOY,airTemp),color='red',size=0.5)+
  scale_x_continuous(lim=c(0,360)) +
  scale_y_continuous("Temperature (C)") +
#    scale_y_continuous(expression(paste("Temperature ( ",degree, "C)", sep = ""))) + #this results in panels not lining up
      geom_vline( aes(xintercept=finalSpringBP), data=springFallBPs[springFallBPs$year %in% 2003:2004 & springFallBPs$site == 'MAUSGS_WEST_BROOK',] ) +
      geom_vline( aes(xintercept=finalFallBP), data=springFallBPs[springFallBPs$year %in% 2003:2004 & springFallBPs$site == 'MAUSGS_WEST_BROOK',] ) +
  theme_bw(base_size=25)+
  theme(axis.title.x = element_blank(),
        axis.text.x = element_blank()) +
  geom_text(aes(label=text),size=10,data=labels1)+
  facet_wrap(~year)

ggsave( file=paste0(graphsDir,'ggRawWB.png'), plot=ggRawWB, dpi=600 , width=6,height=5, units='in', scale=2 )

labels2 <- data.frame(
  dOY = c(345,345),
  movingMean = c(24, 24),
  text = c("c", "d"),
  year = c(2003,2004))

ggWBBPs <- 
ggplot(e[e$riverOrdered == 'WEST BROOK' & e$year %in% 2003:2004,],aes(dOY,movingMean))+
  geom_point()+
 # geom_line(aes(dOY,airTemp),color='red',size=0.5)+
  geom_hline( aes(yintercept=quantileLo), data=springFallBPs[springFallBPs$year %in% 2003:2004 & springFallBPs$site == 'MAUSGS_WEST_BROOK',] ) +
  geom_hline( aes(yintercept=quantileHi), data=springFallBPs[springFallBPs$year %in% 2003:2004 & springFallBPs$site == 'MAUSGS_WEST_BROOK',] ) +
    geom_vline( aes(xintercept=finalSpringBP), data=springFallBPs[springFallBPs$year %in% 2003:2004 & springFallBPs$site == 'MAUSGS_WEST_BROOK',] ) +
      geom_vline( aes(xintercept=finalFallBP), data=springFallBPs[springFallBPs$year %in% 2003:2004 & springFallBPs$site == 'MAUSGS_WEST_BROOK',] ) +
  scale_x_continuous("Day of year",lim=c(0,360)) +
  scale_y_continuous("Temperature Index",lim=c(-25,25) ) +
  theme_bw(base_size=25) +
  theme(strip.text.x = element_blank(),
        strip.background = element_blank()) +
  #ylim(-5,30) +
  geom_text(aes(label=text),size=10,data=labels2)+
  facet_wrap(~year)

rawWB <- arrangeGrob(ggRawWB, ggWBBPs,ncol=1)#, widths=c(5,1.5))  

ggsave( file=paste0(graphsDir,'rawWB.png'), plot=rawWB, dpi=600 , width=6,height=5, units='in', scale=2 )


```

```{r}
# to calculate by hand the proportion of missing data for each site/year combo
springFallBPs$numSyncDays <- springFallBPs$finalFallBP - springFallBPs$finalSpringBP

```

```{r correlations among streams}

forCorr <- tempDataSyncS %>%
  select(date,temp,riverMS) %>%
  spread(riverMS,temp) 

  forCorr$year <- factor(as.numeric(strftime(forCorr$date, "%Y")))
  forCorr$dOY <- (as.numeric(strftime(forCorr$date, "%j")))
  
corr <- cor(forCorr[,2:5],use="complete.obs")
pairs(forCorr[,2:5])

ggplot(forCorr, aes(OL,I)) + 
  geom_point( aes(color=dOY) ) +
  scale_color_gradient(low='blue',high='red') +
  geom_abline(intercept=0,slope=1) +
  facet_wrap(~year)

```


```{r characterize years - hot, cold seasons}
ggplot(tempDataSyncS, aes(dOY,temp)) + geom_point() + facet_grid(year~riverOrdered)

gGam <- gam(temp~s(dOY),data=tempDataSyncS)
g <- data.frame(s=tempDataSyncS$riverMS,y=tempDataSyncS$year,dOY=tempDataSyncS$dOY,t=tempDataSyncS$temp,r=gGam$residuals,f=gGam$fitted.values)

#ggplot(g, aes(dOY,r)) +geom_point() + geom_hline(yintercept=0) + facet_grid(y~s)

ggGam <- 
ggplot(tempDataSyncS,aes(dOYInt,temp)) + 
  geom_point() +
#  geom_smooth(se=F,color='white',size=2.5) +
  scale_x_continuous("Day of Year") +
  scale_y_continuous("Temperature (C)") +
  geom_line(aes(dOY*64.9+196.9,f),size=2.0,color="white",data=g)+
  theme_bw(base_size=25)

ggsave( file=paste0(graphsDir,'ggGam.png'), plot=ggGam, dpi=600 , width=6,height=5, units='in', scale=2 )



gSum <- 
  g %>%
  filter(dOY > (-1.85), !(s == "OS" & y == 2002), !(s == "OS" & y == 2006) ) %>%
  group_by(s,y)%>%
  mutate( rL = lag( r ),
          delta = r - rL,
          cR = cumsum(r),
          wOrA = 'Water')

ggCResids <- 
  ggplot(gSum, aes(dOY*64.9+196.9,cR,color=cR)) + geom_point() + geom_hline(yintercept=0) + 
    scale_color_gradient2(low='blue', high = 'red',mid='darkgrey') +
 #   scale_y_continuous( expression(paste("Cumulative residuals ( ",degree, "C)", sep = ""))) +
    scale_y_continuous("Cumulative residuals (C)") +
    scale_x_continuous("Day of year",breaks=seq(100, 300,100)) +
    theme_bw(base_size=18) +
    theme(legend.position="none") +
    theme(axis.text.x = element_text(size=12,angle=90,vjust=0.5))+
    facet_grid(s~y)

ggsave( file=paste0(graphsDir,'ggCResids.png'), plot=ggCResids, dpi=600 , width=6,height=5, units='in', scale=2 )


ggplot(gSum, aes(dOY,cR,color=factor(s))) + geom_point() + geom_hline(yintercept=0) + facet_wrap(~y)
ggplot(gSum, aes(dOY,delta, color=factor(y))) + geom_line() + geom_hline(yintercept=0) + facet_wrap(~s)

#ggplot(gSum[gSum$s=="MAUSGS_WEST_BROOK",], aes(dOY,cR,color=factor(y))) + geom_point() + geom_hline(yintercept=0) + coord_polar() + facet_wrap(s~y)

gSum %>%
  group_by(s,y) %>%
 # do(acf(.$r,main=unique(paste(.$s,.$y))))
do(plot(abs(fft(.$r)),main=unique(paste(.$s,.$y))))

f=fft(gSum$r)
plot(abs(f))
```

```{r air temp trends}

ggplot(tempDataSyncS, aes(dOY,airTemp)) + geom_point() + facet_grid(year~riverOrdered)

# backtransform the z-scores.
tempDataSyncS$airTempBT <- tempDataSyncS$airTemp*sd(e$airTemp,na.rm=T)+mean(e$airTemp,na.rm=T)
ggplot(tempDataSyncS, aes(dOY,airTempBT)) + geom_point() + facet_grid(year~riverOrdered)


ggGamAT <- 
ggplot(tempDataSyncS,aes(dOYInt,airTempBT)) + geom_point() +geom_smooth(se=F,color='white',size=2.5) +
  scale_x_continuous("Day of Year") +
  scale_y_continuous("Temperature (C)") +
  theme_bw(base_size=25)

ggsave( file=paste0(graphsDir,'ggGamAT.png'), plot=ggGamAT, dpi=600 , width=6,height=5, units='in', scale=2 )

gGamAT <- gam(airTempBT ~ s(dOY),data=tempDataSyncS)
gAT <- data.frame(s=tempDataSyncS$riverMS,y=tempDataSyncS$year,dOY=tempDataSyncS$dOY,t=tempDataSyncS$airTempBT,r=gGamAT$residuals,f=gGamAT$fitted.values)
#ggplot(g, aes(dOY,r)) +geom_point() + geom_hline(yintercept=0) + facet_grid(y~s)

gSumAT <- 
  gAT %>%
  filter( dOY > (-1.85), !(s == "OS" & y == 2002), !(s == "OS" & y == 2006) ) %>%
  group_by(s,y)%>%
  mutate( rL = lag( r ),
          delta = r - rL,
          cR = cumsum(r),
          wOrA = 'Air')

ggCResidsAT <- 
  ggplot(gSumAT, aes(dOY,cR,color=cR)) + geom_point() + geom_hline(yintercept=0) + 
    scale_color_gradient2(low='blue', high = 'red',mid='darkgrey') +
    scale_y_continuous( expression(paste("Cumulative residuals ( ",degree, "C)", sep = ""))) +
    scale_x_continuous("Standardized day of year") +
    theme_bw(base_size=18) +
    theme(legend.position="none") +
    facet_grid(s~y)

ggsave( file=paste0(graphsDir,'ggCResidsAT.png'), plot=ggCResidsAT, dpi=600 , width=6,height=5, units='in', scale=2 )

```
 
```{r combine water and air temp resids}

gSumBoth <- bind_rows(gSum,gSumAT)

ggCResidsBoth <- 
  ggplot(gSumBoth[ gSumBoth$wOrA == 'Water', ], aes(dOY,cR,color=cR)) + geom_point() + geom_hline(yintercept=0) + 
    scale_color_gradient2(low='blue', high = 'red',mid='darkgrey') +
    scale_y_continuous( expression(paste("Cumulative residuals ( ",degree, "C)", sep = ""))) +
    scale_x_continuous("Standardized day of year") +
    geom_line( aes(dOY,cR,color=cR), data=gSumBoth[ gSumBoth$wOrA == 'Air', ] ) +
    theme_bw(base_size=18) +
    theme(legend.position="none") +
    facet_grid(s~y)

ggsave( file=paste0(graphsDir,'ggCResidsBoth.png'), plot=ggCResidsBoth, dpi=600 , width=6,height=5, units='in', scale=2 )

```

```{r air temp, water temp raw data validation graphs}

for(yy in sort(unique(gSum$y))){
gg <-   
ggplot(filter(gSumBoth,wOrA == 'Water',y==yy), aes(dOY,t)) + 
  geom_point(color='red') + 
  geom_line(color='red') +
  geom_line(aes(dOY,f),color='red') +
  
  geom_point( data=filter(gSumBoth,wOrA == 'Air',y==yy), aes(dOY,t),color='blue') +
  geom_line( data=filter(gSumBoth,wOrA == 'Air',y==yy), aes(dOY,t),color='blue') +
  geom_line( data=filter(gSumBoth,wOrA == 'Air',y==yy), aes(dOY,f),color='blue') +
  geom_hline(yintercept=0) + 
  facet_grid(wOrA~s) +
  ggtitle(yy)
print(gg)

#gg2 <- 
#  ggplot(filter(gSum,y==yy), aes(dOY,cR)) + geom_point(color='red') + geom_point  (aes(dOY,r)) + geom_hline(yintercept=0) + facet_grid(~s)+
#  ggtitle(yy)
#print(gg2)
}


```

```{r precip residuals}
# prcp is standardized
# backtransform the z-scores.
tempDataSyncS$prcpBT <- tempDataSyncS$prcp*sd(e$prcp,na.rm=T) + mean(e$prcp,na.rm=T)
ggplot(tempDataSyncS, aes(dOY,prcpBT)) + geom_point() + facet_grid(year~riverOrdered)

#ggplot(tempDataSyncS, aes(dOY,prcp)) + geom_point() + facet_grid(year~riverOrdered)

minDOY <- -1.4 # need this to be smaller than for the temp ones because small differences in BPs among rivers make the rivers have very different cumul resids

# get cumulative precip
cPrcp <- 
  tempDataSyncS %>%
  filter( dOY > (minDOY), !(riverMS == "OS" & year == 2002), !(riverMS == "OS" & year == 2006) ) %>%
  group_by(riverMS,year)%>%
  mutate( cP = cumsum(prcpBT) )

ggCP <- 
ggplot(cPrcp,aes(dOY,cP,color=factor(year))) + geom_point() +
  geom_smooth(se=F,color='white',size=2.5) +
  scale_x_continuous("Day of Year") +
  scale_y_continuous("Precip") +
  theme_bw(base_size=25) 
#  facet_grid(riverMS~year)

ggsave( file=paste0(graphsDir,'ggCP.png'), plot=ggCP, dpi=600 , width=6,height=5, units='in', scale=2 )

gGamP <- gam(cP ~ s(dOY),data=cPrcp)
gP <- data.frame(s=cPrcp$riverMS,y=cPrcp$year,dOY=cPrcp$dOY, t=cPrcp$cP, r=gGamP$residuals, f=gGamP$fitted.values )
#ggplot(gP, aes(dOY,r)) +geom_point() + geom_hline(yintercept=0) + facet_grid(y~s)

# get cumul resids
gSumP <- 
  gP %>%
  filter( dOY > (minDOY), !(s == "OS" & y == 2002), !(s == "OS" & y == 2006) ) %>%
  group_by(s,y)%>%
  mutate( cR = cumsum(r),
          wOrA = 'Prcp')

ggplot(gSumP,aes(dOY,cR,color=factor(s))) + geom_point() +
 # geom_smooth(se=F,color='white',size=2.5) +
  scale_x_continuous("Day of Year") +
  scale_y_continuous("Precip") +
  theme_bw(base_size=25) +
  facet_wrap(~y)

```

```{r combine water and air temp and precip resids}

gSumAll3 <- bind_rows(gSumBoth,gSumP)

gSumAll3$cRPrcpForGraph <- gSumAll3$cR/100

ggCResidsAll3 <- 
  ggplot(gSumAll3[ gSumAll3$wOrA == 'Water', ], aes(dOY,cR,color=cR)) + geom_point() + geom_hline(yintercept=0) + 
    scale_color_gradient2(low='blue', high = 'red',mid='darkgrey') +
    scale_y_continuous( expression(paste("Cumulative residuals ( ",degree, "C)", sep = ""))) +
    scale_x_continuous("Standardized day of year") +
    geom_line( aes(dOY,cR,color=cR), data=gSumAll3[ gSumAll3$wOrA == 'Air', ] ) +
    geom_line( aes(dOY,cRPrcpForGraph),color='black', data=gSumAll3[ gSumAll3$wOrA == 'Prcp', ] ) +
    theme_bw(base_size=18) +
    theme(legend.position="none") +
    facet_grid(s~y)

ggsave( file=paste0(graphsDir,'ggCResidsBoth.png'), plot=ggCResidsBoth, dpi=600 , width=6,height=5, units='in', scale=2 )

```

```{r calc diff btw water and air resids, compare to precip resids}
w <- gSumAll3 %>%
  filter( wOrA == 'Water') %>%
  rename( cRWater=cR )
a <- gSumAll3 %>%
  filter( wOrA == 'Air') %>%
  select(s,y,dOY,cR) %>%
  rename( cRAir=cR )
p <- gSumAll3 %>%
  filter( wOrA == 'Prcp') %>%
  select(s,y,dOY,cR) %>%
  rename( cRPrcp=cR )

wap <- w %>%
  left_join(.,a) %>%
  left_join(.,p) %>%
  mutate( diff = cRWater - cRAir)
  

  ggplot(wap, aes(dOY,diff,color=diff)) + geom_point() + geom_hline(yintercept=0) + 
    scale_color_gradient2(low='blue', high = 'red',mid='darkgrey') +
    scale_y_continuous( expression(paste("Cumulative residuals ( ",degree, "C)", sep = ""))) +
    scale_x_continuous("Standardized day of year") +
    geom_line( aes(dOY,cRPrcp/100),color='black' ) +
    theme_bw(base_size=18) +
    theme(legend.position="none") +
    facet_grid(s~y)
  
  # all this is very interesting, but too complicated for the MS.
```


```{r variance in the aT and wT residuals}

gSumVar <- 
gSumBoth %>%
  group_by(s,y,wOrA) %>%
  summarize( v = var(r) ) %>%
  spread(wOrA,v) %>%
  mutate(ratio = Air/Water)

ggplot( gSumVar, aes( y,ratio )) +
  geom_point(aes(color=s)) +
  geom_line(aes(color=s)) +
  facet_wrap(~s)


```


Model validation

```{r obs v pred}


#observed vs. predicted
obsPred <- 
ggplot(tempDataSyncS[tempDataSyncS$year>=2003,], aes(temp, pred.wb)) +
    geom_point(alpha = 0.3) + 
    geom_abline(intercept=0,slope=1) + 
    scale_x_continuous( expression(paste("Observed temperature ( ",degree, "C)", sep = ""))) +
    scale_y_continuous( expression(paste("Predicted temperature ( ",degree, "C)", sep = ""))) +
    theme_bw(base_size=25) +
    facet_grid(year~riverMS)

ggsave( file=paste0(graphsDir,'obsPred.png'), plot=obsPred, dpi=500, width=9,height=7, units='in', scale=2 )

summary(lm(pred.wb~temp + factor(year)*temp + riverMS*temp,data=tempDataSyncS[tempDataSyncS$year>=2003,]))

```


```{r means tables with rHat}
load( paste0(dataLocalDir,'ggs_wb.RData') )

intersectSeveral <- function(...) { Reduce(intersect, list(...)) } 

# exclude the variables in the grep() statement
# intersect works on pairs. the intersectSeveral function generates the intersection of multiple lists
keepRows <- 
  intersectSeveral(grep("residuals", x=ggs.wb$Parameter,invert=TRUE),
                   grep("stream.mu", x=ggs.wb$Parameter,invert=TRUE),
                   grep("rho.B.year",x=ggs.wb$Parameter,invert=TRUE)
                  )

rHat <- unique( ggs_rhat_temp(ggs.wb[keepRows,], ggs.wb)$data ) # ggs_rhat has problems because using keepRows strips the attributes. function s is saved in wd()/code
rHat$Rhat1 <- ifelse(rHat$Rhat < 1,1,rHat$Rhat)
max(rHat$Rhat1,na.rm=T)

# table of parameter estimates for MS
# decided not to use table in MS?
m <- left_join( rHat[,c('Parameter','Rhat1')], means )

#ggs_caterpillar(ggs.wb[keepRows,])
m$gt2 <- ifelse( m$mean > 2, 1,0 )
m$oP <- factor(m$Parameter,levels=rev( c(
 'B.0[1]','B.0[2]','B.0[3]','B.0[4]','B.0[5]','B.0[6]','B.0[7]','B.0[8]','B.0[9]','B.0[10]','B.0[11]','B.0[12]','mu.year[1]','mu.year[2]','mu.year[3]','mu.year[4]','ar1Mean','ar1SD','ar1[1]','ar1[2]','ar1[3]','ar1[4]','mu.year[1]','mu.year[2]','mu.year[3]','mu.year[4]','sigma'
 ) ), ordered=T)

#m$oP <- 
#plyr::mapvalues(m$oP, from = c('B.0[1]','B.0[2]','B.0[3]','B.0[4]','B.0[5]','B.0[6]','B.0[7]','B.0[8]','B.0[9]','B.0[10]','B.0[11]','B.0[12]'), to = c('B[0]','B[1]','B[2]','B[3]','B[4]','B[5]','B[6]','B[7]','B[8]','B[9]','B[10]','B[11]'))

m$oP <- 
plyr::mapvalues(m$oP, from = c('B.0[1]','B.0[2]','B.0[3]','B.0[4]','B.0[5]','B.0[6]','B.0[7]','B.0[8]','B.0[9]','B.0[10]','B.0[11]','B.0[12]','ar1[1]','ar1[2]','ar1[3]','ar1[4]','mu.year[1]','mu.year[2]','mu.year[3]','mu.year[4]'), to = c('Intercept','T','T(d-1)','T(d-2)','F','T * F','OL','OS','IS','OL * T','OS * T','IS * T','ar1[WB]','ar1[OL]','ar1[OS]','ar1[IL]','0','Linear','Quadratic','Cubic'))


# without B.year[]
# do in two parts to simulate a 'break'
b1 <- 
ggplot(m[grep("B.year",m$Parameter,invert=T) & !is.na(m$oP) & m$oP!="0",], aes(x = mean, y = oP)) + 
  geom_point(size = 3) + 
  geom_segment(aes(x = qLo, xend = qHi, yend = oP), size = 0.5) +
  geom_segment(aes(x = qLo25, xend = qHi75, yend = oP), size = 1.5) +
  theme_bw(base_size=25) +
  scale_x_continuous('        Value',limits=c(-2.7,2)) +
  scale_y_discrete('Parameter') +
  geom_vline(xintercept=0) +
  theme(#axis.text.x = element_text(angle=50),
        strip.text.x = element_blank(),
        strip.background = element_blank(),
        strip.text.y = element_blank(),
        plot.margin= unit(c(1,0.05,1,1), "lines"))
 
b2 <- 
ggplot(m[grep("B.year",m$Parameter,invert=T) & !is.na(m$oP)& m$oP!="0",], aes(x = mean, y = oP)) + 
  geom_point(size = 3) + 
  geom_segment(aes(x = qLo, xend = qHi, yend = oP), size = 0.5) +
  geom_segment(aes(x = qLo25, xend = qHi75, yend = oP), size = 1.5) +
  theme_bw(base_size=25) +
  scale_x_continuous('',limits=c(14.4,15.6), breaks=c(14.5,15.5)) +
  scale_y_discrete('') +
  #facet_grid( ~ gt2, scales='free_x',space='free' ) +
  theme(axis.text.y = element_blank(),
        strip.text.x = element_blank(),
        strip.background = element_blank(),
        strip.text.y = element_blank(),
        plot.margin= unit(c(1,1,1,0), "lines"))
  

betas <- arrangeGrob(b1, b2,ncol=2, widths=c(5,1.5))  

ggsave( file=paste0(graphsDir,'betas2.png'), plot=betas, dpi=600, width=5,height=5, units='in', scale=2 )

data.frame(m) #paste into excel to get ready for appendix table


# B.year[] only
ggplot(m[grep("B.year",m$Parameter,invert=F),], aes(x = mean, y = reorder(Parameter,mean))) + 
  geom_point(size = 3) + 
  geom_segment(aes(x = qLo, xend = qHi, yend = reorder(Parameter,mean)), size = 0.5) +
  geom_segment(aes(x = qLo25, xend = qHi75, yend = reorder(Parameter,mean)), size = 1.5)


```


```{r trend in bPs?}
library(reshape2)

bP <- unique( tempDataSyncS[,c("year","riverMS",'finalSpringBP','finalFallBP')] )
bPM <- melt(bP, id.var=c('year','riverMS'))
bPM$yearCent <- bPM$year-2008
bP
bPM$v2[bPM$variable == 'finalSpringBP'] <- 'Spring break point'
bPM$v2[bPM$variable == 'finalFallBP'] <- 'Fall break point'

ggBP <- 
ggplot( bPM[bPM$year>2001,], aes( year,value, color=riverMS ) ) +
  geom_point(aes(shape=riverMS),size=5)+
  #geom_line(aes( linetype=riverOrdered)) +
  scale_x_continuous('Year') +
  scale_y_continuous( 'Day of year' ) +
  geom_smooth(aes( linetype=riverMS),method='lm',se=F)+
  theme_bw(base_size=25) +
  theme(legend.title=element_blank()) +
  facet_wrap(~v2, ncol=1, scales='free')

ggsave( file=paste0(graphsDir,'ggBP.png'), plot=ggBP, dpi=600, width=5,height=5, units='in', scale=2 )


s1 <- (lm(value~yearCent, bPM[bPM$year>2001 & bPM$variable == 'finalSpringBP',]))
s2 <- (lm(value~yearCent+riverMS, bPM[bPM$year>2001 & bPM$variable == 'finalSpringBP',]))
s3 <- (lm(value~yearCent*riverMS, bPM[bPM$year>2001 & bPM$variable == 'finalSpringBP',]))
AIC(s1,s2,s3)
summary(s1)


f1 <- (lm(value~yearCent, bPM[bPM$year>2001 & bPM$variable == 'finalFallBP',]))
f2 <- (lm(value~yearCent+riverMS, bPM[bPM$year>2001 & bPM$variable == 'finalFallBP',]))
f3 <- (lm(value~yearCent*riverMS, bPM[bPM$year>2001 & bPM$variable == 'finalFallBP',]))
AIC(f1,f2,f3)
summary(f1)

ggBP2 <- 
ggplot( bPM[bPM$year>2001,], aes( year,value ) ) +
  geom_point(aes(shape=riverMS),size=5)+
  #geom_line(aes( linetype=riverOrdered)) +
  scale_x_continuous('Year') +
  scale_y_continuous( 'Day of year' ) +
  geom_smooth(color='black',method='lm',se=F)+
  theme_bw(base_size=25) +
  theme(legend.title=element_blank()) +
  facet_wrap(~v2, ncol=1, scales='free')

ggsave( file=paste0(graphsDir,'ggBP2.png'), plot=ggBP2, dpi=600, width=5,height=5, units='in', scale=2 )

```


```{r predictions of cubic by year}
library(reshape2)
pC <- predCubic(means)

ggCubic <- 
  ggplot(pC$predCubic, aes(dOY*64.9+196.9,pred + 15.09, group=year)) + 
    geom_line(size=1.25) +
    geom_point(aes(maxX*64.9+196.9,maxY + 15.09), color='white',data = pC$bY) +
    scale_x_continuous('Day of year') +
    scale_y_continuous( expression(paste("Temperature ( ",degree, "C)", sep = ""))) +
    theme_bw(base_size=20) 
#    facet_wrap(~year)

ggsave( file=paste0(graphsDir,'ggCubic.png'), plot=ggCubic, dpi=600 , width=6,height=6, units='in' )



ggCubic + geom_point(aes(maxX,maxY), data = pC$bY) +facet_wrap(~year)

# max temp over years
ggMaxTemp <- 
ggplot(pC$bY, aes(year+1998,maxY + 15.09)) +  
  geom_point(size=3) +
  geom_smooth(method="lm",color='black') +
  scale_x_continuous('Year') +
  scale_y_continuous( expression(paste("Temperature ( ",degree, "C)", sep = ""))) +
  theme_bw(base_size=25) +
  theme( legend.position="none")

summary(lm( maxY ~ year, data=pC$bY ))

# day of max temp over years
ggDayOfMaxTemp <- 
ggplot(pC$bY, aes(year+1998,maxX*64.9+196.9)) + 
  geom_point(size=3) +
  geom_smooth(method="lm",color='black') +
  scale_x_continuous('Year') +
  scale_y_continuous( expression(paste("Day of year", sep = ""))) +
  theme_bw(base_size=25) 

summary(lm( maxX ~ year, data=pC$bY ))

ggMaxTempMaxDay <- arrangeGrob(ggMaxTemp,ggDayOfMaxTemp, ncol=1)

  ggsave( file=paste0(graphsDir,'ggMaxTempMaxDay.png'), plot=ggMaxTempMaxDay, dpi=600 , width=5,height=7, units='in', scale=2 )


###########
# look at predictions at specific std values of dOY
bYM <- melt(pC$bY[,c('year','dOYMinus2','dOYMinus1','dOYPlus1','dOYPlus2')],id='year') 
bYM$std <- ifelse(bYM$variable %in% c('dOYMinus1','dOYPlus1'), 1, 2 ) 
bYM$tempAdj <- bYM$value + 18 # 18 for now as a fill-in

predCubicStd <- 
  ggplot(bYM, aes(year+1998,tempAdj,linetype=variable,shape=variable)) +
  geom_point(size=3) +
  geom_smooth(method="lm",color='black') +
  scale_x_continuous('Year') +
  scale_y_continuous( expression(paste("Temperature ( ",degree, "C)", sep = ""))) +
  theme_bw(base_size=25) +
  theme( legend.position="none")# +
#  facet_wrap(~std, scales = "free")

ggsave( file=paste0(graphsDir,'predCubicStd.png'), plot=predCubicStd, dpi=600 , width=6,height=6, units='in' )

# test significance
summary(lm( tempAdj ~ year, data=bYM[bYM$variable == 'dOYMinus2',] ))
summary(lm( tempAdj ~ year, data=bYM[bYM$variable == 'dOYMinus1',] ))
summary(lm( tempAdj ~ year, data=bYM[bYM$variable == 'dOYPlus1',] ))
summary(lm( tempAdj ~ year, data=bYM[bYM$variable == 'dOYPlus2',] ))
```

```{r recalculate RMSE for test/val cross validation}
# need to calculate the rmse for just the validation set and just the test set. separate these by 'keep' the days that were used in the test set.
load( paste0(dataLocalDir,'validation.RData') )  # this is big, so only load when necessary

# melt dOYKeep1
keep <- as.data.frame(dOYKeep1) %>%
  mutate( scenario = 1:nrow(dOYKeep1) ) %>%
  gather( scenario, na.rm=T ) 
  
keep <- keep[,c(1,3)] # can't get this to work in dplyr. dplyr named the first 2 cols'scenario' 

keep$s_d <- paste(keep$scenario,keep$value,sep='_')
predictSummary$s_d <- paste(predictSummary$scenario,predictSummary$dOYInt,sep='_')
#predictSummary$outYear <- predictSummary$year == predictSummary$yearOut 

# test=T if 

# rivsIndex  r   year       dOYInt
#     1     WB   yearOut    s_d in keep$s_d 
#     2     any  yearOut    s_d in keep$s_d 
predictSummary$test <- F
predictSummary$test[predictSummary$rivsIndex == 1 & 
                    predictSummary$r == "WEST BROOK" &  
                    predictSummary$year == predictSummary$yearOut &
                    predictSummary$s_d %!in% keep$s_d  
                   ] <- T
predictSummary$test[predictSummary$rivsIndex == 2 & 
                    predictSummary$year == predictSummary$yearOut &
                    predictSummary$s_d %!in% keep$s_d  
                   ] <- T
table(predictSummary$test)

pS1 <- predictSummary %>%
  #group_by( scenario ) %>%
  #mutate( inKeep = s_d %in% keep$s_d ) %>%
#  group_by( scenario,rivsIndex,year,test ) %>%
  group_by( r,year,yearOut,scenario,rivsIndex,test ) %>%
  summarise( rmseByKeep = rmse( resid ),nDays=n() )
  #mutate( rivOut = paste0(rivsIndex,outYear) )


ggplot(filter(pS1,scenario==28 & test), aes(year,rmseByKeep, color=r,shape=factor(rivsIndex)))+
  geom_point() +
  geom_line() +
  facet_grid(~rivsIndex)

#ggplot(pS1, aes(scenario,nDays,color=factor(rivsIndex)))+geom_point() +facet_grid(test~year)
#ggplot(pS1, aes(scenario,rmseByKeep,color=factor(rivsIndex)))+geom_point() +facet_grid(test~year)

# get base-case (training) rmse for merging
# just broken down by river and year
meanTrainData <- pS1 %>%
  filter( !test ) %>%
  rename( rmseTrain = rmseByKeep ) %>%
  group_by(r,year) %>%
  summarize( rmseByRYearForTrain = mean(rmseTrain))
  
#pS <- left_join(pS1,meanTrainData, by=c("scenario", "rivsIndex", "year")) %>%
pS <- left_join(pS1,meanTrainData, by=c('r','year')) %>%
  mutate( diff = rmseByKeep - rmseByRYearForTrain )

ggplot(filter(pS,scenario==28 & year==yearOut ), aes(year,diff, color=r,shape=factor(rivsIndex)))+
  geom_point() +
  geom_line() +
  facet_grid(~rivsIndex)

# Table for MS
# summary across rivers, all years
filter(pS,scenario==28 ) %>%
  group_by(test,rivsIndex) %>%
  summarize( mDiff=mean(diff),sDiff=sd(diff),
             mRmse=mean(rmseByKeep),sRmse=sd(rmseByKeep),n=n())


# summary across rivers for non-test data
filter(pS1,scenario==28 & !test ) %>%
  #group_by(rivsIndex) %>%
  ungroup() %>%
  summarize( #mDiff=mean(diff),sDiff=sd(diff),
             mRmse=mean(rmseByKeep),sRmse=sd(rmseByKeep),n=n())



s28 <- filter(pS,scenario==28 ) %>%
  #mutate(o=outYear*1) %>%
  unite(rO,test,rivsIndex,sep="") %>%
 # select(-outYear) %>%
  spread(rO,rmseByKeep) 
 
ggplot(s28, aes(year,FALSE1)) +
  geom_point() +
  geom_point(aes(year,FALSE2), color='red') +
  geom_line(aes(year,FALSE2), color='red') +
  geom_point(aes(year,TRUE1), color='blue') +
  geom_line(aes(year,TRUE1), color='blue') +
  geom_point(aes(year,TRUE2), color='darkgreen') +
  geom_line(aes(year,TRUE2), color='darkgreen')



# get means across years
pS2 <- pS %>%
  #filter(outYear) %>%
  group_by(scenario,rivsIndex,test) %>%
  summarize(rmseMean=mean(rmseByKeep),rmseSD=sd(rmseByKeep),n=n(), 
            meanDiff=mean(diff),sdDiff=sd(diff)) 

ggplot( filter(pS, scenario %in% 14:22 & test), aes(scenario,rmseByKeep,color=factor(rivsIndex) ) )+
  geom_point(  ) +
  geom_smooth(method='lm'
#              , formula=y~x+I(x^2)
              ,se=F)+#,color='black') +
  scale_x_continuous('Number of days included in estimation') +
  scale_y_continuous('Difference in RMSE from base case') +
  theme_bw(base_size=25) +
  theme(legend.position="none")    +facet_wrap(~year)
```  
  
  
```{r plots for cross validation, randomly left out points}

load( file = paste0(dataLocalDir,'crossValidationPropOut.RData'))

ggplot(summ,aes(propOut,mean,color=keep)) + geom_point()

cvPropOut <- 
ggplot(filter(rmseByOutAllCV, propOut<0.81),aes(propOut,rmse,shape=keep)) + 
  geom_point(size=3.5) +
  scale_x_continuous('Proportion of data left out') +
  scale_y_continuous('Root mean square error') +
  theme_bw(base_size=25) +
  theme(legend.position="none") 

ggsave( file=paste0(graphsDir,'cvPropOut.png'), plot=cvPropOut, dpi=600 , width=7,height=5, units='in', scale=2 )

summary(lm(rmse~propOut,data=filter(rmseByOutAllCV, propOut<0.81,keep)))

# means and sd of test data
mean(filter(rmseByOutAllCV, propOut<0.81,!keep)$rmse)
sd(filter(rmseByOutAllCV, propOut<0.81,!keep)$rmse)
```

```{r plots for cross-validation}
load( paste0(dataLocalDir, "r.RData"))

#ggplot( r[r$incompleteYear & !is.na(r$riverOrdered),], aes(incompleteIndex,diff)) + geom_point()+facet_wrap(~rivsIndex)

####
#cross-valiation for MS



meansByS <-   
r %>%
  filter( incompleteYear,numD > 20, rivsIndex==1 & riverOrdered=='WEST BROOK' | rivsIndex==2 ) %>%
  group_by( rivsIndex,scenario ) %>%
    summarise( mean = mean( diff, na.rm=T ),
               sd   =   sd( diff, na.rm=T ),
               n    =    n(  ),
             dOYDiff = min(dOYDiff),
             numD = mean( numD, na.rm=T ),
             numDSD = sd( numD, na.rm=T ),
                  s = unique(s),
             meanMinDOY = mean( minDOY, na.rm=T ),
             meanPropD = mean( propD, na.rm=T )) %>%
  filter( !is.na( s )
         # , s=='middle' 
         ) %>%
  mutate( dOYDiff = ifelse( dOYDiff == -Inf,0, dOYDiff ),
          lo = mean - sd,
          hi = mean + sd) 


ggplot( r[r$propD>0,] ,aes(numD,diff, color=rivsIndex
          #     ,shape=factor(s)
               )) +
  geom_point()+#aes(color=factor(s)) ) +
  facet_grid(scenario~riverOrdered)

l <- lm(mean~meanPropD*factor(rivsIndex)+0, data=meansByS[meansByS$s=='middle',])
#q <- lm(mean~numD*factor(rivsIndex)*I(numD^2)*factor(rivsIndex), data=meansByS[meansByS$s=='middle',])
#AIC(l,q) # q is better

#just all left out
summary(lm(mean~meanPropD, data=meansByS[meansByS$s=='middle' & meansByS$rivsIndex == 1,])) 

#just WB left out
summary(lm(mean~meanPropD, data=meansByS[meansByS$s=='middle' & meansByS$rivsIndex == 2,]))


ggRMSEDOY <- 
ggplot( meansByS[meansByS$s=='middle',] ,aes(meanPropD,mean
               ,shape=factor(rivsIndex), , linetype=factor(rivsIndex)
               )) +
  geom_point( size=4 ) +
 # geom_errorbar( aes(numD,ymin=lo,ymax=hi)) +
  geom_smooth(method='lm'
              , formula=y~x #+ I(x^2)
              ,se=F,color='black') +
 # geom_point( aes(dOYDiff,mean), data=meansByS[meansByS$s == 'earlyChunk',], shape=1,size=4 ) +
#  geom_point( aes(dOYDiff,mean), data=meansByS[meansByS$s == 'lateChunk',], shape=5,size=3  ) +
  scale_x_continuous('Proportion of days included in estimation') +
  scale_y_continuous('') +
  theme_bw(base_size=25) +
  theme(legend.position="none") 

ggsave( file=paste0(graphsDir,'ggRMSEDOY.png'), plot=ggRMSEDOY, dpi=600 , width=6,height=5, units='in', scale=2 )

# break it up by river and year
r$rY <- paste0(r$riverOrdered,r$year)

ggMiddleScenarios <- 
ggplot( na.omit(r[r$s %in% c('0days','middle') 
                  & r$riverOrdered == "WEST BROOK"
                  ,]) ,aes(numD,diff
          #     ,shape=factor(s)
                ,color=factor(rivsIndex)
               )) +
  geom_point(  ) +
  geom_smooth(method='lm'
              #, formula=y~x+I(x^2)
              ,se=F)+#,color='black') +
  scale_x_continuous('Number of days included in estimation') +
  scale_y_continuous('Difference in RMSE from base case') +
  theme_bw(base_size=25) +
  theme(legend.position="none") +
  facet_wrap(~year)

ggsave( file=paste0(graphsDir,'ggMiddleScenarios.png'), plot=ggMiddleScenarios, dpi=600 , width=8,height=5, units='in', scale=2 )


#6=2002, 11=2007, 13=2009
for(i in 13:21){print(ggObsPred[[16+1*i]])}


```

```{r look at 30day patterns}
# want to excude incomplete years - they skew the means
meansByS2 <-   
r %>%
  filter( incompleteYear,numD > 20, rivsIndex==1 & riverOrdered=='WEST BROOK' | rivsIndex==2, year %in% c(1997:2007,2010,2011) ) %>%
  group_by( rivsIndex,scenario ) %>%
    summarise( mean = mean( diff, na.rm=T ),
               sd   =   sd( diff, na.rm=T ),
               n    =    n(  ),
             dOYDiff = min(dOYDiff),
             numD = mean( numD, na.rm=T ),
             numDSD = sd( numD, na.rm=T ),
                  s = unique(s),
             meanMinDOY = mean( minDOY, na.rm=T )) %>%
  filter( !is.na( s )
         # , s=='middle' 
         ) %>%
  mutate( dOYDiff = ifelse( dOYDiff == -Inf,0, dOYDiff ),
          lo = mean - sd,
          hi = mean + sd) 

ggRMSE30Days <- 
ggplot( meansByS2[meansByS2$s=='30days' & meansByS2$meanMinDOY < 280 & meansByS2$meanMinDOY > 75,] ,aes(meanMinDOY,mean
               ,shape=factor(rivsIndex), linetype=factor(rivsIndex)
               )) +
  geom_point( size=4 ) +
 # geom_errorbar( aes(meanMinDOY,ymin=lo,ymax=hi)) +
  geom_smooth(method='lm'
              , formula=y~x #+ I(x^2)
              ,se=F,color='black') +
 # geom_point( aes(dOYDiff,mean), data=meansByS[meansByS$s == 'earlyChunk',], shape=1,size=4 ) +
#  geom_point( aes(dOYDiff,mean), data=meansByS[meansByS$s == 'lateChunk',], shape=5,size=3  ) +
  scale_x_continuous('Starting day (day of year)') +
  scale_y_continuous('') +
  theme_bw(base_size=25) +
  theme(legend.position="none") 

ggsave( file=paste0(graphsDir,'ggRMSEDOY.png'), plot=ggRMSEDOY, dpi=600 , width=6,height=5, units='in', scale=2 )

ggRMSEBoth <- arrangeGrob(ggRMSEDOY,ggRMSE30Days, left= textGrob("Difference in RMSE from base case", gp=gpar(fontsize=25),rot=90),ncol=1) #, widths=c(5,1.5))  

ggsave( file=paste0(graphsDir,'ggRMSEBoth.png'), plot=ggRMSEBoth, dpi=600 , width=5,height=5, units='in', scale=2 )


gg30DayScenarios <- 
ggplot( na.omit(r[r$s %in% c('30days') 
                  & r$riverOrdered == "WEST BROOK"
                  & r$numD == 30
                  ,]) ,aes(70+scenario*20,diff  
                ,shape=factor(rivsIndex)
                ,color=factor(rivsIndex)
               )) +
  geom_point(  ) +
  geom_smooth(method='lm'
              ,formula=y~x+I(x^2) #+I(x^3)
              ,se=F)+#,color='black') +
  scale_x_continuous('Starting day (day of year)') +
  scale_y_continuous('Difference in RMSE from base case') +
  theme_bw(base_size=25) +
  theme(legend.position="none") +
  facet_wrap(~year)

ggsave( file=paste0(graphsDir,'gg30DayScenarios.png'), plot=gg30DayScenarios, dpi=600 , width=8,height=5, units='in', scale=2 )



```

```{r look at cumulative rmse}
# cumulResid <- 
# predictSummary %>%
#   group_by(scenario,yearOut) %>%
#   mutate( c = cumsum(resid) )
# 
# ggplot(cumulResid, aes(dOYInt,c)) +
#   geom_point()
```
